+++ xyselect/example_test.go
@@ -1,6 +1,7 @@
 package xyselect_test
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/xybor/xyplatform/xyselect"
@@ -19,8 +20,20 @@ func ExampleE() {
 	var _, v, _ = eselector.Select(false)
 	fmt.Println(v)
 
+	var _, _, err = eselector.Select(false)
+	if errors.Is(err, xyselect.ClosedChannelError) {
+		fmt.Println(err)
+	}
+
+	_, _, err = eselector.Select(false)
+	if errors.Is(err, xyselect.ExhaustedError) {
+		fmt.Println(err)
+	}
+
 	// Output:
 	// 10
+	// ClosedChannelError: channel closed
+	// ExhaustedError: selector is exhausted
 }
 
 func ExampleR() {